Export `IsUnion` type
