+++ source/single-key-object.d.ts
@@ -1,6 +1,6 @@
 import type {IsEmptyObject} from './empty-object.js';
 import type {If} from './if.js';
-import type {IsUnion} from './internal/index.d.ts';
+import type {IsUnion} from './is-union.d.ts';
 
 /**
 Create a type that only accepts an object with a single key.
